requirements.txt
python-telegram-bot
sqlalchemy
requests
